+++ lenstronomy/LensModel/lens_model.py
@@ -3,6 +3,7 @@ __author__ = 'sibirrer'
 from lenstronomy.LensModel.single_plane import SinglePlane
 from lenstronomy.LensModel.multi_plane import MultiPlane
 from lenstronomy.Cosmo.lens_cosmo import LensCosmo
+from astropy.cosmology import default_cosmology
 
 
 class LensModel(object):
@@ -11,7 +12,7 @@ class LensModel(object):
     """
 
     def __init__(self, lens_model_list, z_lens=None, z_source=None, lens_redshift_list=None, cosmo=None,
-                 multi_plane=False, numerical_alpha_class=None, observed_convention_index=False):
+                 multi_plane=False, numerical_alpha_class=None, observed_convention_index=None, z_source_convention=None):
         """
 
         :param lens_model_list: list of strings with lens model names
@@ -27,23 +28,33 @@ class LensModel(object):
         (see documentation in Profiles/numerical_alpha)
         :param observed_convention_index: a list of lens indexes that correspond to observed positions on the sky, not
         physical positions
+        :param z_source_convention: float, redshift of a source to define the reduced deflection angles of the lens
+        models. If None, 'z_source' is used.
         """
         self.lens_model_list = lens_model_list
         self.z_lens = z_lens
+        if z_source_convention is None:
+            z_source_convention = z_source
         self.z_source = z_source
+        self._z_source_convention = z_source_convention
         self.redshift_list = lens_redshift_list
+
+        if cosmo is None:
+            cosmo = default_cosmology.get()
         self.cosmo = cosmo
         self.multi_plane = multi_plane
         if multi_plane is True:
             if z_source is None:
                 raise ValueError('z_source needs to be set for multi-plane lens modelling.')
+
             self.lens_model = MultiPlane(z_source, lens_model_list, lens_redshift_list, cosmo=cosmo,
                                          numerical_alpha_class=numerical_alpha_class,
-                                         observed_convention_index=observed_convention_index)
+                                         observed_convention_index=observed_convention_index,
+                                         z_source_convention=z_source_convention)
         else:
             self.lens_model = SinglePlane(lens_model_list, numerical_alpha_class=numerical_alpha_class)
         if z_lens is not None and z_source is not None:
-            self._lensCosmo = LensCosmo(z_lens, z_source, cosmo=self.cosmo)
+            self._lensCosmo = LensCosmo(z_lens, z_source, cosmo=cosmo)
 
     def ray_shooting(self, x, y, kwargs, k=None):
         """
@@ -81,11 +92,11 @@ class LensModel(object):
         :param x_image: image position
         :param y_image: image position
         :param kwargs_lens: lens model parameter keyword argument list
-        :return:
+        :return: arrival time of image positions in units of days
         """
-        if hasattr(self.lens_model, 'arrival_time'):
+        try:
             arrival_time = self.lens_model.arrival_time(x_image, y_image, kwargs_lens)
-        else:
+        except:
             x_source, y_source = self.lens_model.ray_shooting(x_image, y_image, kwargs_lens)
             fermat_pot = self.lens_model.fermat_potential(x_image, y_image, x_source, y_source, kwargs_lens)
             if not hasattr(self, '_lensCosmo'):