+++ HISTORY.rst
@@ -105,3 +105,11 @@ History
 +++++++++++++++++
 * psf_construction simplified
 * parameter handling for catalogue modelling improved
+
+0.9.0 (2019-7-06)
++++++++++++++++++
+* faster fft convolutions
+* re-design of multi-plane lensing module
+* re-design of plotting module
+* nested samplers implemented
+* Workflow module with added features